Add files via upload
Documentation provided by the Google Cybersecurity certificate program from Coursera. This information helped me conduct an internal security audit for "Botium Toys". The security audit checklist is the documentation used to conduct the test and the other two files explain the issues the company is facing. I used the template to conduct the audit and added my own evaluation.